refactor(types): export interface/type 공통화 및 타입 import 구조 개선
- UserData, UserWithId, Task, friendsProps 등 export interface/type을 src/lib/type/interface.ts로 통합
- 각 컴포넌트 및 백엔드 코드에서 공통 타입을 import하여 일관성 및 재사용성 향상
- ToDoDashboad, ToDoTask 등에서 Task 타입 import 경로를 interface.ts로 변경
- 기존 개별 파일 내 타입 선언 제거 및 import 구조 정리
