+++ src/components/src/main/HomeClient.tsx
@@ -3,10 +3,10 @@
 import { useFriendSearch } from "@/contexts/FriendSearchContext";
 import FriendSearchResults from "@/components/src/friends/FriendSearchResults";
 import { FriendList } from "@/components/src/main/FriendsList";
-import ToDoListsDashboard, { Task } from "@/components/src/main/ToDoDashboad";
+import ToDoListsDashboard from "@/components/src/main/ToDoDashboad";
 import LogOut from "@/components/src/main/LogOutButton";
-import { UserWithId } from "@/lib/backend/write-firebase";
 import { friendsProps } from "@/app/(main)/page";
+import { Task, UserWithId } from "@/lib/type/interface";
 
 interface HomeClientProps {
   initialTasks: Task[];
